requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
tensorflow-cpu